Added recalculation proposal integration tests with feature flag
