+++ tests/Nhs.Appointments.Api.Integration/Scenarios/AvailabilityCalculations/RecalculationProposal_Disabled.feature
@@ -1,0 +1,5 @@
+Feature: Availability Recalculation Proposal Disabled
+
+  Scenario: Recalculate availability change proposal
+    When I request recalculation proposal endpoint
+    Then the call should fail with 404
