+++ tests/Nhs.Appointments.Api.Integration/Scenarios/BaseFeatureSteps.cs
@@ -1,21 +1,25 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Net.Http;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
 using AutoMapper;
 using FluentAssertions;
 using Gherkin.Ast;
 using Microsoft.Azure.Cosmos;
 using Microsoft.Azure.Cosmos.Linq;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using Nhs.Appointments.Api.Auth;
+using Nhs.Appointments.Api.Json;
+using Nhs.Appointments.Api.Models;
 using Nhs.Appointments.Core;
 using Nhs.Appointments.Core.Json;
 using Nhs.Appointments.Persistance;
 using Nhs.Appointments.Persistance.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Net;
+using System.Net.Http;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
 using Xunit;
 using Xunit.Gherkin.Quick;
 using Feature = Xunit.Gherkin.Quick.Feature;
@@ -39,6 +43,8 @@ public abstract partial class BaseFeatureSteps : Feature
     protected readonly HttpClient Http;
     protected readonly Mapper Mapper;
     protected List<BookingDocument> MadeBookings = new List<BookingDocument>();
+    protected readonly Dictionary<int, string> _bookingReferences = new();
+    protected HttpResponseMessage _response { get; set; }
 
     public BaseFeatureSteps()
     {
@@ -645,6 +651,83 @@ public abstract partial class BaseFeatureSteps : Feature
         await AssertAvailabilityStatusByReference(customId, expectedStatus, false);
     }
 
+    [When(@"I create the following availability")]
+    public async Task CreateAvailability(DataTable dataTable)
+    {
+        foreach (var row in dataTable.Rows.Skip(1))
+        {
+            var cells = row.Cells.ToList();
+
+            var date = cells.ElementAt(0).Value;
+            var from = cells.ElementAt(1).Value;
+            var until = cells.ElementAt(2).Value;
+            var slotLength = cells.ElementAt(3).Value;
+            var capacity = cells.ElementAt(4).Value;
+            var services = cells.ElementAt(5).Value;
+
+            var payload = new
+            {
+                date = ParseNaturalLanguageDateOnly(date).ToString("yyyy-MM-dd"),
+                site = GetSiteId(),
+                sessions = new[]
+                {
+                    new
+                    {
+                        from,
+                        until,
+                        slotLength = int.Parse(slotLength),
+                        capacity = int.Parse(capacity),
+                        services = services.Split(',').Select(s => s.Trim()).ToArray()
+                    }
+                },
+                mode = "additive"
+            };
+
+            _response = await Http.PostAsJsonAsync("http://localhost:7071/api/availability", payload);
+            _response.StatusCode.Should().Be(HttpStatusCode.OK);
+        }
+    }
+
+    [Then("I make the following bookings")]
+    public async Task MakeBookings(DataTable dataTable)
+    {
+        var bookingIndex = 0;
+        foreach (var row in dataTable.Rows.Skip(1))
+        {
+            var cells = row.Cells.ToList();
+            var date = cells.ElementAt(0).Value;
+            var time = cells.ElementAt(1).Value;
+            var duration = cells.ElementAt(2).Value;
+            var service = cells.ElementAt(3).Value;
+
+            object payload = new
+            {
+                from = DateTime.ParseExact(
+                    $"{ParseNaturalLanguageDateOnly(date):yyyy-MM-dd} {time}",
+                    "yyyy-MM-dd HH:mm", null).ToString("yyyy-MM-dd HH:mm"),
+                duration,
+                service,
+                site = GetSiteId(),
+                kind = "booked",
+                attendeeDetails = new
+                {
+                    nhsNumber = NhsNumber,
+                    firstName = "John",
+                    lastName = "Doe",
+                    dateOfBirth = "1987-03-13"
+                }
+            };
+            _response = await Http.PostAsJsonAsync("http://localhost:7071/api/booking", payload);
+            _response.StatusCode.Should().Be(HttpStatusCode.OK);
+
+            var result =
+                JsonConvert.DeserializeObject<MakeBookingResponse>(await _response.Content.ReadAsStringAsync());
+            var bookingReference = result.BookingReference;
+
+            _bookingReferences[bookingIndex] = bookingReference;
+            bookingIndex += 1;
+        }
+    }
     [And("there are no sessions for '(.+)'")]
     public async Task AssertSessionNoLongerExists(string dateString)
     {
